validarLogin
Validé el login e hice todos los cambios pertinentes en los respectivos archivos

+++ components/funcionTraerPokemones.php
@@ -1,19 +1,8 @@
 <?php
-//Establecer parametros para la conexion a la base de datos
-$serverName = "localhost";
-$username = "root";
-$password = "";
-$database = "pokedex";
-// Establecer la conexión con la base de datos
-$conexion = mysqli_connect($serverName, $username, $password, $database);
-
-// Verificar si la conexión es exitosa
-if (!$conexion) {
-    die("La conexión falló: " . mysqli_connect_error());
-}
+include_once("conexion.php");
 
 // Construir la consulta SQL
-$sql = "SELECT * FROM Pokemones";
+$sql = "SELECT * FROM pokemones";
 
 // Ejecutar la consulta y obtener los resultados
 $resultado = $conexion->query($sql);
@@ -26,7 +15,12 @@ $resultado = $conexion->query($sql);
         <th> Tipo</th>
         <th> N°</th>
         <th> Nombre</th>
-        <th> Acciones</th>
+        <?php
+        if(!empty($_COOKIE['seguridad']) && !empty($_SESSION["nombreUsuario"])){
+            if($_COOKIE['seguridad']==$hash){
+                echo("<th> Acciones</th>");
+            }
+        }?>
     </tr>
     </thead>
 
@@ -46,16 +40,19 @@ $resultado = $conexion->query($sql);
 
             echo "<td class='tipo'>" . $elemento["nombre"] . "</td>";
 
-            echo "<td class='acciones'>";
+            if(!empty($_COOKIE['seguridad']) && !empty($_SESSION["nombreUsuario"])){
+                if($_COOKIE['seguridad']==$hash){
+                    echo "<td class='acciones'>";
 
-            echo "<a href='components/updatePokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='update fa-solid fa-pen-to-square'></i></a>";
+                    echo "<a href='components/updatePokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='update fa-solid fa-pen-to-square'></i></a>";
 
-            echo "<a href='components/deletePokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='delete fa-solid fa-trash'></i></a>";
+                    echo "<a href='components/deletePokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='delete fa-solid fa-trash'></i></a>";
 
-            echo "<a href='components/Pokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='fa-solid fa-eye'></i></a>";
+                    echo "<a href='components/Pokemon.php?id=" . $elemento["IDPokemon"] . "' ><i class='fa-solid fa-eye'></i></a>";
 
-            echo '</td>';
-
+                    echo '</td>';
+                        }
+            }
             echo ' </tr>';
         }
